+++ src/hooks/query/queryClient.ts
@@ -1,0 +1,11 @@
+import { QueryClient } from 'react-query';
+
+export const queryClient = new QueryClient({
+  defaultOptions: {
+    queries: {
+      retry: 0,
+      staleTime: 10 * 600 * 1000,
+      enabled: true,
+    },
+  },
+});
