+++ src/hooks/query/useGetQuery.tsx
@@ -11,45 +11,28 @@ import {
   UserAPI,
 } from '@api/api';
 import { UserState } from '@store/user';
-import { queryKeys } from '.';
+import { queryKeys } from './queryKeys';
 import { IMentoringListRequestOptions } from 'types/api.types';
 
 export const useGetUserAPI = () => {
   const [user, setUser] = useRecoilState(UserState);
 
-  return useQuery(
-    queryKeys.user,
-    () => UserAPI.getUserInfo().catch((err) => err),
-    {
-      // 브라우저 focus 됐을 때 재시작?
-      retry: false,
-      refetchOnWindowFocus: false,
-      // 자동으로 가져오는 옵션
-      enabled: true,
-      // 캐시 타임
-      staleTime: 10 * 600 * 1000,
-      onSuccess: (res) => {
-        if (res?.data?.body.data) {
-          setUser(res.data.body.data);
-        }
-      },
+  return useQuery(queryKeys.user, () => UserAPI.getUserInfo(), {
+    refetchOnWindowFocus: false,
+    onSuccess: (res) => {
+      if (res?.data?.body.data) {
+        setUser(res.data.body.data);
+      }
     },
-  );
+    onError: () => {
+      console.log('에러');
+    },
+  });
 };
 
 export function useGetPostByPostId(postId: number) {
-  return useQuery(
-    queryKeys.post(postId),
-    () => PostAPI.getPostByPostId(postId),
-    {
-      // 브라우저 focus 됐을 때 재시작?
-      retry: false,
-      refetchOnWindowFocus: true,
-      // 자동으로 가져오는 옵션
-      enabled: true,
-      // 캐시 타임
-      staleTime: 10 * 600 * 1000,
-    },
+  return useQuery(queryKeys.post(postId), () =>
+    PostAPI.getPostByPostId(postId),
   );
 }
 
@@ -80,13 +63,8 @@ export function useGetSearchPostList({
         sortOption,
       }).then((response) => response.data.body.data),
     {
-      // 브라우저 focus 됐을 때 재시작?
-      retry: false,
       refetchOnWindowFocus: false,
-      // 자동으로 가져오는 옵션
       enabled: !!keyword,
-      // 캐시 타임
-      staleTime: 10 * 600 * 1000,
     },
   );
 }
@@ -96,26 +74,14 @@ export function useGetCategoryList() {
     queryKeys.categories,
     () => CategoryAPI.getCategoryList().catch((err) => err),
     {
-      // 브라우저 focus 됐을 때 재시작?
-      retry: false,
       refetchOnWindowFocus: false,
-      // 자동으로 가져오는 옵션
-      enabled: true,
-      // 캐시 타임
-      staleTime: 10 * 600 * 1000,
     },
   );
 }
 
 export function useGetRecommendList() {
   return useQuery(queryKeys.recommend, () => PostAPI.recommendPostList(), {
-    // 브라우저 focus 됐을 때 재시작?
-    retry: false,
     refetchOnWindowFocus: false,
-    // 자동으로 가져오는 옵션
-    enabled: true,
-    // 캐시 타임
-    staleTime: 10 * 600 * 1000,
   });
 }
 
@@ -124,39 +90,20 @@ export function useGetLikeofPost(postId: number) {
     queryKeys.likePost(postId),
     () => LikeAPI.getIsLikeInfo(postId),
     {
-      // 브라우저 focus 됐을 때 재시작?
-      retry: false,
       refetchOnWindowFocus: false,
-      // 자동으로 가져오는 옵션
-      enabled: true,
-      // 캐시 타임
-      staleTime: 10 * 600 * 1000,
-      onError: () => {},
     },
   );
 }
 
 export function useGetAlarm() {
   return useQuery(queryKeys.alarms, () => AlarmAPI.getAlarm(), {
-    retry: false,
-    refetchOnWindowFocus: true,
-    enabled: true,
     staleTime: 0,
   });
 }
 
 export function useGetApplicant(postId: number) {
-  return useQuery(
-    queryKeys.applicant(postId),
-    () => ApplicationAPI.getApplicant(postId),
-    {
-      retry: false,
-      refetchOnWindowFocus: true,
-      // 자동으로 가져오는 옵션
-      enabled: true,
-      // 캐시 타임
-      staleTime: 10 * 600 * 1000,
-    },
+  return useQuery(queryKeys.applicant(postId), () =>
+    ApplicationAPI.getApplicant(postId),
   );
 }
